moves System roles state and openHandbook to mainComponent
